Fix broken compulaton in linux
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3331 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/samson/stream/BlockLookupList.cpp
@@ -139,7 +139,7 @@ namespace samson {
             if( key_size > key_size2 )
                 return 1;
             
-            for (int i = 0 ; i < key_size ; i++ )
+            for (size_t i = 0 ; i < key_size ; i++ )
             {
                 
                 if( key[i] < key2[i] )
